Split benchmark receipt contracts into core microcrate
